.history/requirements_20241229190243.txt
aiogram==3.12.0
fastapi==0.100.0
uvicorn==0.23.0
requests==2.31.0
python-dotenv==1.0.0
